requirements.txt
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
isodate==0.6.1
mpegdash==0.4.0
oauthlib==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ratelimit==2.2.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
spotipy==2.23.0
tidalapi==0.7.5
typing_extensions==4.11.0
urllib3==1.26.6
